Pass on README and doc/
Copy @Nyan11's interesting page from GH wiki about Bloc, Toplo
and StyleSheets. This way this and future documentation .md pages
can be read not only from the web browser but also from the
Documentation Browser in the Pharo's  World Menu. The README
now has the direct link and explains it.

This makes it the README and doc/ more homogeneus with other
projets in the ecosystem such as Bloc, Album and Alexandrie.